Move editor entity spawning into its own function.

+++ src/game/default/g_types.h
@@ -718,8 +718,8 @@ typedef struct {
   g_client_t *clients; // [sv_max_clients]
 
   bool ai_loaded; // whether the AI is loaded or not
-  uint8_t ai_fill_slots; // total number of empty slots the AI should fill
-  uint8_t ai_left_to_spawn; // the number of AI bots that we're waiting to spawn in
+  int32_t ai_fill_slots; // total number of empty slots the AI should fill
+  int32_t ai_left_to_spawn; // the number of AI bots that we're waiting to spawn in
 } g_game_t;
 
 extern g_game_t g_game;
